Working controller

+++ service/aws/instance.go
@@ -2,6 +2,7 @@ package aws
 
 import (
 	"fmt"
+	"time"
 
 	"encoding/json"
 	"sync"
@@ -23,7 +24,8 @@ type AWSMachine struct {
 }
 
 var (
-	infrastructureMutex sync.Mutex
+	infrastructureMutex            sync.Mutex
+	checkForExistsAfterCreateCount = 25
 )
 
 func New(region string, profile string) (service.MachineMutation, error) {
@@ -48,12 +50,34 @@ func (a *AWSMachine) Create(machine *clusterv1.Machine) (string, error) {
 	infrastructureMutex.Lock()
 	defer infrastructureMutex.Unlock()
 	pc := getProviderConfig(machine.Spec.ProviderConfig)
-	spl := strings.Split(pc.Name, ".")
+	spl := strings.Split(machine.Name, ".")
 	clusterName := spl[0]
-	logger.Info(pc.ServerPool.Image)
 	if pc.ServerPool.Image == "" {
 		return "", nil
 	}
+	logger.Debug("AMI: %s\n", pc.ServerPool.Image)
+	logger.Debug("KeyPair: %s\n", clusterName)
+
+	// Calculate Security Groups
+	var sgs []*string
+	for _, firewall := range pc.ServerPool.Firewalls {
+		sgs = append(sgs, &firewall.Identifier)
+	}
+
+	// Calculate Subnet IDs and Create NIC
+	var nics []*ec2.InstanceNetworkInterfaceSpecification
+	index := 0
+	for _, subnet := range pc.ServerPool.Subnets {
+		nic := &ec2.InstanceNetworkInterfaceSpecification{
+			SubnetId:                 aws.String(subnet.Identifier),
+			AssociatePublicIpAddress: aws.Bool(true),
+			DeviceIndex:              aws.Int64(int64(index)),
+			Groups:                   sgs,
+		}
+		index++
+		nics = append(nics, nic)
+	}
+
 	input := &ec2.RunInstancesInput{
 		ImageId:      aws.String(pc.ServerPool.Image),
 		InstanceType: aws.String(pc.ServerPool.Size),
@@ -61,6 +85,11 @@ func (a *AWSMachine) Create(machine *clusterv1.Machine) (string, error) {
 		MaxCount:     aws.Int64(1),
 		KeyName:      aws.String(clusterName),
 		UserData:     aws.String(string(pc.ServerPool.GeneratedNodeUserData)),
+		//SecurityGroups:    sgs,
+		NetworkInterfaces: nics,
+		IamInstanceProfile: &ec2.IamInstanceProfileSpecification{
+			Name: aws.String(pc.ServerPool.InstanceProfile.Name),
+		},
 	}
 	output, err := a.EC2.RunInstances(input)
 	if err != nil {
@@ -69,8 +98,12 @@ func (a *AWSMachine) Create(machine *clusterv1.Machine) (string, error) {
 	tagInput := &ec2.CreateTagsInput{
 		Tags: []*ec2.Tag{
 			{
-				Key:   aws.String("KubicornCluster"),
-				Value: aws.String(pc.Name),
+				Key:   aws.String("KubernetesCluster"),
+				Value: aws.String(clusterName),
+			},
+			{
+				Key:   aws.String("Name"),
+				Value: aws.String(machine.Name),
 			},
 		},
 		Resources: []*string{
@@ -79,11 +112,25 @@ func (a *AWSMachine) Create(machine *clusterv1.Machine) (string, error) {
 	}
 	_, err = a.EC2.CreateTags(tagInput)
 	if err != nil {
-		defer a.Destroy(pc.Name)
+		defer a.Destroy(machine.Name)
 		logger.Warning("Unable to tag instance: destroying: %v", err)
 		return "", fmt.Errorf("Unable to tag instance: %v", err)
 	}
-	logger.Info("Created instance: %s", output.Instances[0].InstanceId)
+
+	// --- Hang until instance registers
+	i := 0
+	for {
+		if a.Exists(machine.Name) {
+			return *output.Instances[0].InstanceId, nil
+		}
+		time.Sleep(time.Second * 5)
+		i++
+		if i == checkForExistsAfterCreateCount {
+			return "", fmt.Errorf("Unable to detect instance after create and minimual checks expired")
+		}
+		logger.Info("Waiting for machine to register [%s]...", machine.Name)
+	}
+	logger.Always("Created instance: %s", *output.Instances[0].InstanceId)
 	return *output.Instances[0].InstanceId, nil
 }
 
@@ -92,8 +139,9 @@ func (a *AWSMachine) Exists(name string) bool {
 		logger.Info("Empty name")
 		return true
 	}
-	infrastructureMutex.Lock()
-	defer infrastructureMutex.Unlock()
+	logger.Info("Query for instance name [%s]", name)
+	//infrastructureMutex.Lock()
+	//defer infrastructureMutex.Unlock()
 	input := &ec2.DescribeInstancesInput{
 		Filters: []*ec2.Filter{
 			{
@@ -111,36 +159,82 @@ func (a *AWSMachine) Exists(name string) bool {
 		return true
 	}
 	if len(output.Reservations) < 1 {
-		return true
+		logger.Always("Machine [%s] NOT found", name)
+		return false
 	}
 
-	// TODO we assume that we have reservations
-	if len(output.Reservations[0].Instances) > 1 {
-		return true
+	logger.Debug("Reservation count: %d", len(output.Reservations))
+	for _, reservation := range output.Reservations {
+		instances := reservation.Instances
+		logger.Debug("Instance count: %d", len(instances))
+		for _, instance := range instances {
+			if *instance.State.Name != "running" {
+				logger.Debug("Instance not `running` state is [%s]", *instance.State.Name)
+				continue
+			}
+			tags := instance.Tags
+			for _, tag := range tags {
+				k := *tag.Key
+				v := *tag.Value
+				if k == "Name" {
+					if v == name {
+						logger.Always("Machine [%s] found", name)
+						return true
+					}
+					logger.Debug("Instance found but not matched [%s][%s]", name, v)
+				}
+			}
+		}
 	}
+
+	logger.Always("Machine [%s] NOT found", name)
 	return false
 }
 
-func (a *AWSMachine) Destroy(id string) error {
+func (a *AWSMachine) Destroy(name string) error {
+	logger.Always("Destroying instance: %s", name)
 	infrastructureMutex.Lock()
 	defer infrastructureMutex.Unlock()
-	input := &ec2.TerminateInstancesInput{
-		InstanceIds: []*string{
-			&id,
+	input := &ec2.DescribeInstancesInput{
+		Filters: []*ec2.Filter{
+			{
+				Name: aws.String("tag:Name"),
+				Values: []*string{
+					aws.String(name),
+				},
+			},
 		},
 	}
-	_, err := a.EC2.TerminateInstances(input)
+	output, err := a.EC2.DescribeInstances(input)
 	if err != nil {
-		return fmt.Errorf("Unable to destroy instance [%s]: %v", id, err)
+		return fmt.Errorf("Unable to destroy instance: %v", err)
 	}
+	for _, reservation := range output.Reservations {
+		for _, instance := range reservation.Instances {
+			if *instance.State.Name == "running" {
+				input := &ec2.TerminateInstancesInput{
+					InstanceIds: []*string{
+						instance.InstanceId,
+					},
+				}
+				_, err := a.EC2.TerminateInstances(input)
+				if err != nil {
+					return fmt.Errorf("Unable to destroy instance [%s]: %v", name, err)
+				}
+				logger.Always("Terminated instance: %s", name)
+			}
+		}
+	}
+
 	return nil
 }
 
 func (a *AWSMachine) ListIDs(name string) ([]string, error) {
+	logger.Always("List IDs for cluster [%s]", name)
 	input := &ec2.DescribeInstancesInput{
 		Filters: []*ec2.Filter{
 			{
-				Name: aws.String("tag:KubicornCluster"),
+				Name: aws.String("tag:KubernetesCluster"),
 				Values: []*string{
 					aws.String(name),
 				},
@@ -151,11 +245,26 @@ func (a *AWSMachine) ListIDs(name string) ([]string, error) {
 	if err != nil {
 		return []string{}, fmt.Errorf("Unable to list ids: %v", err)
 	}
-	var ids []string
-	for _, instance := range output.Reservations[0].Instances {
-		ids = append(ids, *instance.InstanceId)
+	var names []string
+	if len(output.Reservations) == 0 {
+		return names, nil
 	}
-	return ids, nil
+	for _, reservation := range output.Reservations {
+		for _, instance := range reservation.Instances {
+			if *instance.State.Name != "running" {
+				continue
+			}
+			tags := instance.Tags
+			for _, tag := range tags {
+				k := *tag.Key
+				v := *tag.Value
+				if k == "Name" {
+					names = append(names, v)
+				}
+			}
+		}
+	}
+	return names, nil
 }
 
 func getProviderConfig(providerConfig string) *cluster.MachineProviderConfig {
